+++ src/components/index.js
@@ -5,6 +5,7 @@ export { default as Chips } from './Chips';
 export { default as ContainerFeedback } from './ContainerFeedback';
 export { default as ContentList } from './ContentList';
 export { default as CopyToClipboard } from './CopyToClipboard';
+export { default as CrossTab } from './CrossTab';
 export { default as Deck } from './Deck';
 export { default as Divisor } from './Divisor';
 export { default as FilterBar } from './FilterBar';