+++ CHANGELOG.md
@@ -63,3 +63,5 @@ Toda a biblioteca foi reorganizada para comportar os componentes compartilhados 
 -   [dep] Add dependencias redux-persist e redux-persist-pouchdb
 
 # 2.4.1
+
+-   [components] Add CrossTab